moved generateboilerplate to utils
